Reorganize components and remove duplicate styling
