+++ src/components/auth/StartLogo.tsx
